Mostly Style Changes with a timer reduction
Reduced the timer from 3 minutes to 2 minutes.

Added a message at the game over screen that tells the player why they lost.

Reduced some padding on the title and game over screen to center the text more.

Changed the size of the title text and game over text.

Changed the color of the play and game over background.